Move mobile entry drawer to separate file
