+++ src/components/ui/GlobalStyle.js
@@ -31,6 +31,7 @@ const theme = {
   orange: '#ffa41b',
   transparentOrange: '#ffedd1',
   transparentBlue: '#d9e6f3',
+  transparentPink: '#fbe2fb',
   blue: '#4183c4',
   green: '#73cd7c',
   red: '#ff8188',
