+++ src/utils/api.ts
@@ -91,135 +91,9 @@ export const getTypeById = (
   id: paths['/types/{id}']['get']['parameters']['path']['id'],
 ) => handleResponse(instance.get(`/types/${id}`))
 
-// export const getReviews = (
-//   locationId: paths['/locations/{id}/reviews']['get']['parameters']['path']['id'],
-// ) => handleResponse(instance.get(`/locations/${locationId}/reviews`))
-
-export const getReviews = () =>
-  // locationId: paths['/locations/{id}/reviews']['get']['parameters']['path']['id'],
-  Promise.resolve([
-    {
-      id: 2745,
-      location_id: 3383,
-      user_id: null,
-      author: 'Ethan',
-      created_at: '2013-12-13T06:09:25.536Z',
-      updated_at: '2013-12-30T19:12:03.511Z',
-      observed_on: null,
-      comment: null,
-      fruiting: null,
-      quality_rating: 3,
-      yield_rating: 2,
-      photos: [
-        {
-          thumb:
-            'https://s3-us-west-2.amazonaws.com/fallingfruit-production/observations/photos/000/002/745/thumb/open-uri20131213-3992-1szjh9k.jpg',
-          medium:
-            'https://s3-us-west-2.amazonaws.com/fallingfruit-production/observations/photos/000/002/745/medium/open-uri20131213-3992-1szjh9k.jpg',
-          original:
-            'https://s3-us-west-2.amazonaws.com/fallingfruit-production/observations/photos/000/002/745/original/open-uri20131213-3992-1szjh9k.jpg',
-        },
-      ],
-    },
-    {
-      id: 6179,
-      location_id: 3383,
-      user_id: 1,
-      author: 'Arman',
-      created_at: '2014-06-21T19:11:21.382Z',
-      updated_at: '2014-06-21T19:11:21.382Z',
-      observed_on: '2014-06-19',
-      comment: 'Here are some pictures I took of the location',
-      fruiting: 0,
-      quality_rating: 1,
-      yield_rating: 2,
-      photos: [
-        {
-          thumb:
-            'https://jeffreytang.com/static/177133048a37d1510361c98ad9aabbc1/c9c6a/020%20Chicago%203G8A5759.webp',
-          medium:
-            'https://jeffreytang.com/static/177133048a37d1510361c98ad9aabbc1/c9c6a/020%20Chicago%203G8A5759.webp',
-          original:
-            'https://jeffreytang.com/static/177133048a37d1510361c98ad9aabbc1/c9c6a/020%20Chicago%203G8A5759.webp',
-        },
-        {
-          thumb:
-            'https://jeffreytang.com/static/0e748c63f0b645282075430007645e9c/c9c6a/010%20Whistler%203G8A3467.webp',
-          medium:
-            'https://jeffreytang.com/static/0e748c63f0b645282075430007645e9c/c9c6a/010%20Whistler%203G8A3467.webp',
-          original:
-            'https://jeffreytang.com/static/0e748c63f0b645282075430007645e9c/c9c6a/010%20Whistler%203G8A3467.webp',
-        },
-        {
-          thumb:
-            'https://jeffreytang.com/static/d32706b879c4d864dc5876afc821a65e/c0e95/030%20Carmel-by-the-Sea%203G8A6620.webp',
-          medium:
-            'https://jeffreytang.com/static/d32706b879c4d864dc5876afc821a65e/c0e95/030%20Carmel-by-the-Sea%203G8A6620.webp',
-          original:
-            'https://jeffreytang.com/static/d32706b879c4d864dc5876afc821a65e/c0e95/030%20Carmel-by-the-Sea%203G8A6620.webp',
-        },
-      ],
-    },
-    {
-      id: 6179,
-      location_id: 3383,
-      user_id: 1,
-      author: 'Arman',
-      created_at: '2014-06-21T19:11:21.382Z',
-      updated_at: '2014-06-21T19:11:21.382Z',
-      observed_on: '2014-06-19',
-      comment: 'Very yummy olives. I put them on my pizza',
-      fruiting: 2,
-      quality_rating: 3,
-      yield_rating: 3,
-      photos: [
-        {
-          thumb:
-            'https://jeffreytang.com/static/408167ba1df5de2564646ed620cc359a/0d27e/IMG_5714.webp',
-          medium:
-            'https://jeffreytang.com/static/408167ba1df5de2564646ed620cc359a/0d27e/IMG_5714.webp',
-          original:
-            'https://jeffreytang.com/static/408167ba1df5de2564646ed620cc359a/0d27e/IMG_5714.webp',
-        },
-        {
-          thumb:
-            'https://jeffreytang.com/static/f1911fbf05c836f7eb7ebcbad19b9de3/0d27e/IMG_5667.webp',
-          medium:
-            'https://jeffreytang.com/static/f1911fbf05c836f7eb7ebcbad19b9de3/0d27e/IMG_5667.webp',
-          original:
-            'https://jeffreytang.com/static/f1911fbf05c836f7eb7ebcbad19b9de3/0d27e/IMG_5667.webp',
-        },
-      ],
-    },
-    {
-      id: 47763,
-      location_id: 3383,
-      user_id: null,
-      author: null,
-      created_at: '2021-05-20T12:39:28.313Z',
-      updated_at: '2021-05-20T12:39:28.313Z',
-      observed_on: null,
-      comment: 'Ethan is reviewing',
-      fruiting: null,
-      quality_rating: null,
-      yield_rating: null,
-      photos: [],
-    },
-    {
-      id: 47764,
-      location_id: 3383,
-      user_id: null,
-      author: null,
-      created_at: '2021-05-20T12:39:32.267Z',
-      updated_at: '2021-05-20T12:39:32.267Z',
-      observed_on: null,
-      comment: 'Ethan is reviewing',
-      fruiting: null,
-      quality_rating: null,
-      yield_rating: null,
-      photos: [],
-    },
-  ])
+export const getReviews = (
+  locationId: paths['/locations/{id}/reviews']['get']['parameters']['path']['id'],
+) => handleResponse(instance.get(`/locations/${locationId}/reviews`))
 
 export const getReviewById = (
   id: paths['/reviews/{id}']['get']['parameters']['path']['id'],
